rename files for better alignment
