+++ test/extensions/transport_sockets/tls/cert_validator/spiffe/spiffe_validator_integration_test.cc
@@ -1,4 +1,4 @@
-#include "spiffe_cert_validator_integration_test.h"
+#include "spiffe_validator_integration_test.h"
 
 #include <memory>
